requirements.txt
streamlit
pandas
pytesseract
Pillow
requests
beautifulsoup4
